requirements.txt
fastapi
uvicorn
motor
prometheus_fastapi_instrumentator
pydantic
python-dotenv
httpx
PyJWT
aioboto3
redis
pgpy
pycrypto
pycryptodome
requests
eth-account
web3